% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CalculateHUM}
\alias{CalculateHUM}
\title{The main function to calculate HUM with specified method}
\usage{
CalculateHUM(
  in_emat_list,
  para_exact = F,
  para_type = c(1, 3),
  para_precision_type = 3,
  para_kappa2 = 25,
  para_prob_bound = 0.05,
  para_R,
  para_random_seed = 1,
  eta = "auto",
  estse = 0
)
}
\description{
We offer four method here.
ACC: All combinations checking
AP: Adaptive precision with parameter kappa (default as 5)
FP: Fixed precision with parameter precision_type (i.e. the required precision as 10^(-precision_type))
}
\examples{
dt1 = gene_data(50,4,0.7,1)
CalculateHUM(dt1,para_type = c(1,3),para_exact = T)

}
